Microservices Arq - Updates .env, removes values, use token placeholders - Updates services, adds OperationService - Updates OperationController, replaces OperationRequest w Service - Adds Operation Service, removes Request
