+++ .env
@@ -23,6 +23,10 @@ APP_SECRET=0727be28ffe7c3daf1701c889c51521d
 # IMPORTANT: You MUST configure your server version, either here or in config/packages/doctrine.yaml
 #
 # DATABASE_URL="sqlite:///%kernel.project_dir%/var/data.db"
-DATABASE_URL="mysql://cllbgsah_operations_user:dbpass@mi3-ss48.a2hosting.com:3306/cllbgsah_operations?serverVersion=5.7"
+# NUNCA PONER VALORES SENSIBLES (DE CONEXION, USERS, PASSES ETC) EN EL .env
+# SE USAN PLACE HOLDERS CON ESTE FORMATO [*PARAMETRO*]
+# LOS VALORES REALES SE USAN EN EL .env.local, QUE ESTÁ IGNORADO
+# EL CHISTE ES NUNCA TENER LOS VALORES REALES EN EL REPO
+DATABASE_URL="mysql://[*DB_USER*]:[*DB_PASS*]@[*DB_HOST*]:[*DB_PORT*]/[*DB_NAME*]?serverVersion=5.7"
 # DATABASE_URL="postgresql://db_user:db_password@127.0.0.1:5432/db_name?serverVersion=13&charset=utf8"
 ###< doctrine/doctrine-bundle ###
